avant finalisation des RMD
