+++ doc/courses/r/03_Graphics.Rmd
@@ -161,10 +161,10 @@ mymodel = Model_createFromParam(ECov_SPHERICAL(), ranges = c(50, 30), angles = c
 simtub(NULL, mygrid, mymodel, nbsimu=2, nbtuba=1000, namconv=NamingConvention("Simu"))
 
 # Representing the two figures on the same pa
-p1 = plot.init(asp=1) + plot.raster(mygrid, name="Simu.1",
+p1 = plot.init(asp=1) + plot.raster(mygrid, name="Simu.S1",
                                     flagLegend=TRUE, legendName="Colors #1")
 p1 = p1 + plot.decoration(title="Simulation 1", xlab="Easting", ylab="Northing")
-p2 = plot.init(asp=1) + plot.raster(mygrid, name="Simu.2", flagLegend=FALSE)
+p2 = plot.init(asp=1) + plot.raster(mygrid, name="Simu.S2", flagLegend=FALSE)
 p2 = p2 + plot.decoration(title="Simulation 2", xlab="Easting", ylab="")
 p = ggarrange(p1, p2, ncol = 2, nrow = 1, align="hv")
 plot.end(p)
