+++ doc/demo/r/Tuto_CalcPostProp.Rmd
@@ -129,7 +129,7 @@ for (i in 1:N) {
 
 # conversion into uniform thickness
 for (i in 1:N) {
-  val = matrix(surfaces$getColumns(names = paste0("Y_", i , ".", 1:nbsim[i]),
+  val = matrix(surfaces$getColumns(names = paste0("Y_", i , ".S", 1:nbsim[i]),
                                    useSel = FALSE),
                nrow = surfaces$getNSample(useSel = FALSE), ncol = nbsim[i])
   val = h_mean[i] * pnorm(val)
