+++ include/Neigh/ANeigh.hpp
@@ -101,8 +101,11 @@ namespace gstlrn
 
     virtual bool getFlagContinuous() const { return false; }
 
-    static ANeigh*
-      createDefaultNeighborhood(ANeigh* neigh, Db* dbin, Id maxNumber = 500);
+    static ANeigh* createDefaultNeighborhood(
+      ANeigh* neigh,
+      const Db* dbin = nullptr,
+      const Db* dbout = nullptr,
+      Id maxNumber = 500);
 
     void displayDebug(VectorInt& ranks) const;
     void select(Id iech_out, VectorInt& ranks);
