+++ swig/swig_exp.i
@@ -412,6 +412,7 @@
 
 %include Simulation/Simulations.hpp
 %include Simulation/ACalcSimulation.hpp
+%include Simulation/ACalcSimuModel.hpp
 %include Simulation/CalcSimuTurningBands.hpp
 %include Simulation/TurningBandDirection.hpp
 %include Simulation/TurningBandOperate.hpp
