+++ swig/swig_inc.i
@@ -378,6 +378,7 @@
 
   #include "Simulation/Simulations.hpp"
   #include "Simulation/ACalcSimulation.hpp"
+  #include "Simulation/ACalcSimuModel.hpp"
   #include "Simulation/CalcSimuTurningBands.hpp"
   #include "Simulation/TurningBandDirection.hpp"
   #include "Simulation/TurningBandOperate.hpp"
