+++ tests/cpp/output/test_SimGaussian.ref
@@ -58,17 +58,17 @@ SimuTB.Data.2.2       0.028      -0.940       0.013      -0.036       1.000     
 SimuTB.Data.2.3      -0.040      -0.079      -0.944       0.035       0.036       1.000
 
 <----- Simulation using Spectral Method ----->
-                   Number     Minimum     Maximum        Mean    St. Dev.    Variance
-SimuSPT.V1.S1   10000.000      87.630     113.597     100.239       3.908      15.270
-SimuSPT.V1.S2   10000.000      83.558     113.877     100.360       3.684      13.573
-SimuSPT.V1.S3   10000.000      87.101     111.853     100.152       3.409      11.623
-SimuSPT.V2.S1   10000.000     -11.031      10.006      -0.283       3.319      11.018
-SimuSPT.V2.S2   10000.000     -11.190      14.614      -0.028       3.355      11.254
-SimuSPT.V2.S3   10000.000     -10.304      10.947      -0.296       2.788       7.772
-               *SPT.V1.S1  *SPT.V1.S2  *SPT.V1.S3  *SPT.V2.S1  *SPT.V2.S2  *SPT.V2.S3
-SimuSPT.V1.S1       1.000      -0.051      -0.071      -0.944       0.055       0.070
-SimuSPT.V1.S2      -0.051       1.000       0.058       0.058      -0.954      -0.060
-SimuSPT.V1.S3      -0.071       0.058       1.000       0.093      -0.064      -0.930
-SimuSPT.V2.S1      -0.944       0.058       0.093       1.000      -0.052      -0.072
-SimuSPT.V2.S2       0.055      -0.954      -0.064      -0.052       1.000       0.075
-SimuSPT.V2.S3       0.070      -0.060      -0.930      -0.072       0.075       1.000
+                      Number     Minimum     Maximum        Mean    St. Dev.    Variance
+SimuSPT.Data.1.1   10000.000      88.069     113.727     100.680       3.712      13.778
+SimuSPT.Data.1.2   10000.000      87.486     111.400      99.778       3.367      11.340
+SimuSPT.Data.1.3   10000.000      85.158     113.243      99.462       3.634      13.208
+SimuSPT.Data.2.1   10000.000     -11.623      10.628      -0.681       3.127       9.777
+SimuSPT.Data.2.2   10000.000      -9.588      10.317       0.292       2.846       8.102
+SimuSPT.Data.2.3   10000.000     -10.576      11.281       0.440       2.838       8.052
+                  *.Data.1.1  *.Data.1.2  *.Data.1.3  *.Data.2.1  *.Data.2.2  *.Data.2.3
+SimuSPT.Data.1.1       1.000      -0.048      -0.031      -0.930       0.040       0.004
+SimuSPT.Data.1.2      -0.048       1.000      -0.016       0.059      -0.916       0.032
+SimuSPT.Data.1.3      -0.031      -0.016       1.000       0.038      -0.001      -0.940
+SimuSPT.Data.2.1      -0.930       0.059       0.038       1.000      -0.048      -0.020
+SimuSPT.Data.2.2       0.040      -0.916      -0.001      -0.048       1.000      -0.003
+SimuSPT.Data.2.3       0.004       0.032      -0.940      -0.020      -0.003       1.000
