+++ tests/data/PGS1/Result.ref
@@ -89,50 +89,50 @@ Data Base Statistics
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       1.502
- Standard Deviation  =       0.616
- Variance            =       0.379
+ Mean value          =       1.266
+ Standard Deviation  =       0.569
+ Variance            =       0.324
 2 - Name Facies.2 - Locator facies2
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       1.814
- Standard Deviation  =       0.941
- Variance            =       0.886
+ Mean value          =       1.734
+ Standard Deviation  =       0.964
+ Variance            =       0.929
 3 - Name Facies.3 - Locator facies3
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       2.179
- Standard Deviation  =       0.919
- Variance            =       0.845
+ Mean value          =       2.272
+ Standard Deviation  =       0.843
+ Variance            =       0.711
 4 - Name Facies.4 - Locator facies4
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       2.576
- Standard Deviation  =       0.804
- Variance            =       0.647
+ Mean value          =       2.563
+ Standard Deviation  =       0.821
+ Variance            =       0.675
 5 - Name Facies.5 - Locator facies5
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       1.896
- Standard Deviation  =       0.779
- Variance            =       0.607
+ Mean value          =       1.765
+ Standard Deviation  =       0.832
+ Variance            =       0.693
 
 Data Base Contents
 ------------------
                Facies.1   Facies.2   Facies.3   Facies.4   Facies.5
-      [  0,]      2.000      1.000      1.000      3.000      3.000
-      [  1,]      2.000      3.000      1.000      3.000      3.000
-      [  2,]      2.000      3.000      1.000      3.000      3.000
-      [  3,]      2.000      3.000      1.000      3.000      3.000
-      [  4,]      2.000      3.000      1.000      3.000      3.000
-      [  5,]      2.000      3.000      1.000      3.000      3.000
-      [  6,]      2.000      3.000      1.000      3.000      3.000
+      [  0,]      1.000      1.000      1.000      3.000      3.000
+      [  1,]      1.000      3.000      1.000      3.000      3.000
+      [  2,]      1.000      3.000      1.000      3.000      3.000
+      [  3,]      1.000      3.000      1.000      3.000      3.000
+      [  4,]      1.000      3.000      1.000      3.000      3.000
+      [  5,]      1.000      3.000      1.000      3.000      3.000
+      [  6,]      1.000      3.000      1.000      3.000      3.000
 (Nrows=7[from 10000])
