+++ tests/data/PGS2/Result.ref
@@ -108,33 +108,33 @@ Data Base Statistics
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       1.540
- Standard Deviation  =       0.542
- Variance            =       0.294
+ Mean value          =       1.432
+ Standard Deviation  =       0.540
+ Variance            =       0.291
 2 - Name Facies.2 - Locator facies2
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       1.946
- Standard Deviation  =       0.807
- Variance            =       0.651
+ Mean value          =       1.724
+ Standard Deviation  =       0.894
+ Variance            =       0.800
 3 - Name Facies.3 - Locator facies3
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       2.256
- Standard Deviation  =       0.928
- Variance            =       0.861
+ Mean value          =       2.285
+ Standard Deviation  =       0.926
+ Variance            =       0.858
 4 - Name Facies.4 - Locator facies4
  Nb of data          =       10000
  Nb of active values =       10000
  Minimum value       =       1.000
  Maximum value       =       3.000
- Mean value          =       1.659
- Standard Deviation  =       0.866
- Variance            =       0.751
+ Mean value          =       1.477
+ Standard Deviation  =       0.817
+ Variance            =       0.667
 
 Data Base Contents
 ------------------
@@ -144,6 +144,6 @@ Data Base Contents
       [  2,]      2.000      1.000      3.000      1.000
       [  3,]      2.000      1.000      3.000      3.000
       [  4,]      2.000      3.000      3.000      1.000
-      [  5,]      2.000      3.000      2.000      1.000
-      [  6,]      2.000      3.000      2.000      1.000
+      [  5,]      1.000      3.000      1.000      1.000
+      [  6,]      1.000      3.000      1.000      1.000
 (Nrows=7[from 10000])
